Moved makeElementDraggable to utils, cleanup
